refact: fix some static issues.

+++ src/smo/utils/scaling.py
@@ -165,7 +165,7 @@ def decide_replicas(
     problem.solve(solver=cp.HIGHS)
 
     if problem.status == cp.OPTIMAL:
-        solution = [int(round(r.value)) for r in r_current]
+        solution = [round(r.value) for r in r_current]
         return solution
     else:
         return None
